+++ src/params/float.rs
@@ -269,6 +269,8 @@ impl FloatParam {
     /// Build a new [`FloatParam`]. Use the other associated functions to modify the behavior of the
     /// parameter.
     pub fn new(name: impl Into<String>, default: f32, range: FloatRange) -> Self {
+        range.assert_validity();
+
         Self {
             value: AtomicF32::new(default),
             normalized_value: AtomicF32::new(range.normalize(default)),